% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Artist_Track_Valence_Energy_Visual.r
\name{artist_track_valence_energy_visual}
\alias{artist_track_valence_energy_visual}
\title{Create a valenece and energy plot based on an artist's tracks}
\usage{
artist_track_valence_energy_visual(
  artist_ids,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{artist_ids}{\itemize{
\item A vector of Spotify artist ids
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A valence and energy plot which plots the top tracks from multiple artists
}
\description{
Create a valenece and energy plot based on an artist's tracks
}
\examples{
\dontrun{
 artist_track_valence_energy_visual("0du5cEVh5yTK9QJze8zA0C")
 artist_track_valence_energy_visual(c("0du5cEVh5yTK9QJze8zA0C", "73sIBHcqh3Z3NyqHKZ7FOL"))
}
}
