% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Artist_Album_Valence_Energy_Visual.r
\name{create_artist_album_valence_energy_visual}
\alias{create_artist_album_valence_energy_visual}
\title{Create a valenece and energy plot based on an artist's albums}
\usage{
create_artist_album_valence_energy_visual(
  artist_id,
  limit = 5,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{artist_id}{\itemize{
\item A single Spotify artist id
}}

\item{limit}{\itemize{
\item Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A valence and energy plot taking in a set number of artist's albums
}
\description{
Create a valenece and energy plot based on an artist's albums
}
\examples{
\dontrun{
 create_artist_album_valence_energy_visual("0du5cEVh5yTK9QJze8zA0C")
 create_artist_album_valence_energy_visual("0du5cEVh5yTK9QJze8zA0C", limit = 5)
}
}
