% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Average_Artists_Radar_Chart.r
\name{create_average_artists_radar_chart}
\alias{create_average_artists_radar_chart}
\title{Create a radar chart of artist features}
\usage{
create_average_artists_radar_chart(
  artists,
  vars = c(),
  colors = c(),
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{artists}{\itemize{
\item A vector of Spotify artist ids
}}

\item{vars}{\itemize{
\item A vector of variables returned from get_artist_summary()
}}

\item{colors}{\itemize{
\item A vector of colors
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A radar chart displaying valence, energy, and speechiness, along with any other inputed variables
}
\description{
Create a radar chart of artist features
}
\examples{
\dontrun{
 create_average_artists_radar_chart(artists = c("5me0Irg2ANcsgc93uaYrpb", "7hJcb9fa4alzcOq3EaNPoG", "1ZwdS5xdxEREPySFridCfh", "7B4hKK0S9QYnaoqa9OuwgX", "1P8IfcNKwrkQP5xJWuhaOC"), vars = c("acousticness", "danceability"), colors = c("#2596BE", "#BC22BF", "#eb4034", "#31cc46", "#8c880f"))
}
}
