% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Violin_Plot_Album_Danceability.r
\name{create_violin_plot_album_danceability}
\alias{create_violin_plot_album_danceability}
\title{Create a violin plot based on an artist's albums}
\usage{
create_violin_plot_album_danceability(
  artist_id,
  limit = 7,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{artist_id}{\itemize{
\item A single Spotify artist id
}}

\item{limit}{\itemize{
\item Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 7
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A violin plot displaying a specific number of an artist's albums' danceability
}
\description{
Create a violin plot based on an artist's albums
}
\examples{
\dontrun{
 create_violin_plot_album_danceability("5K4W6rqBFWDnAN6FQUkS6x")
 create_violin_plot_album_danceability("5K4W6rqBFWDnAN6FQUkS6x", limit = 5)
}
}
