% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Violin_Plot_Artist_Danceability.r
\name{create_violin_plot_artist_danceability}
\alias{create_violin_plot_artist_danceability}
\title{Create a violin plot based on an artist's tracks}
\usage{
create_violin_plot_artist_danceability(
  artist_ids,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{artist_ids}{\itemize{
\item A vector of Spotify artist ids
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A violin plot displaying a number of artist's tracks' danceability
}
\description{
Create a violin plot based on an artist's tracks
}
\examples{
\dontrun{
 create_violin_plot_artist_danceability("0du5cEVh5yTK9QJze8zA0C")
}
}
