% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Genre_Valence_Energy_Visual.r
\name{genre_valence_energy_visual}
\alias{genre_valence_energy_visual}
\title{Create a valenece and energy plot based on genre data}
\usage{
genre_valence_energy_visual(
  genres,
  sample_sizes = 20,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{genres}{\itemize{
\item A vector of Spotify genres
}}

\item{sample_sizes}{\itemize{
\item Number of tracks in each genre to gather data from.  Valid if between 1 and 50.  Defaults to 20
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A valence and energy plot taking in a set number of genres
}
\description{
Create a valenece and energy plot based on genre data
}
\examples{
\dontrun{
 genre_valence_energy_visual(c("hip-hop", "rap", "jazz"))
 genre_valence_energy_visual("rock", sample_sizes = 22)
}
}
